Refactor logging and database handling for modularity
Introduce modular database initialization and logging via `db_init.py` and `db_logging.py`, consolidating repetitive code across modules. Replaced inlined logic in `advanced_knowledge_agent.py` with reusable functions, and improved error handling in relevant sections. Updated `App.jsx` with safer default handling for API responses and configurations. Added a lightweight `calculator` helper for testing purposes.
